QR Code Reader Station/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pyzbar==0.1.9
pytz==2025.2
qrcode==8.2
Pillow==11.3.0 
qrcode[pil]==8.2
pyserial==3.5
evdev==1.9.2